test(workflow-operator): move source-descriptor test stubs out of scanned package
PythonCodeRawInvalidTextSpec uses PythonReflectionUtils.scanCandidates
with acceptPackages = Seq("org.apache.texera.amber.operator") to find
every concrete PythonOperatorDescriptor implementation, then instantiates
each one and runs py_compile against its generatePythonCode() output.

The new PythonSourceOperatorDescriptorSpec's inner StubPythonSource was
landing inside that package prefix and being picked up by the scanner.
Two failure modes followed:

  - Instantiation: Scala-inner-classes have a constructor that requires
    the outer reference, so getDeclaredConstructor() (no-arg) throws.
  - py_compile: the stub returns 'yield {...}' which is a runtime-only
    Python statement and is not a valid top-level module.

Move both stubs (StubSource and StubPythonSource) into a sibling test
package, org.apache.texera.amber.testsupport.source, which is outside the
scanner's accept-prefix. The specs now import the stubs instead of
declaring them as inner classes. No production-code changes; all 9 spec
tests still pass; PythonCodeRawInvalidTextSpec now sees 117/117 OK
(was 117/118).
